Move bug same path

+++ src/Storage/Device.php
@@ -2,6 +2,8 @@
 
 namespace Utopia\Storage;
 
+use Exception;
+
 abstract class Device
 {
     /**
@@ -90,7 +92,7 @@ abstract class Device
      * @param  array  $metadata
      * @return int
      *
-     * @throws \Exception
+     * @throws Exception
      */
     abstract public function upload(string $source, string $path, int $chunk = 1, int $chunks = 1, array &$metadata = []): int;
 
@@ -108,7 +110,7 @@ abstract class Device
      * @param  array  $metadata
      * @return int
      *
-     * @throws \Exception
+     * @throws Exception
      */
     abstract public function uploadData(string $data, string $path, string $contentType, int $chunk = 1, int $chunks = 1, array &$metadata = []): int;
 
@@ -162,6 +164,10 @@ abstract class Device
      */
     public function move(string $source, string $target): bool
     {
+        if ($source === $target) {
+            return false;
+        }
+
         if ($this->transfer($source, $target, $this)) {
             return $this->delete($source);
         }
